feat : 내정보 수정에서 닉네임 중복검사 로직 추가 #27

+++ src/main/java/com/sendback/domain/user/service/UserService.java
@@ -99,6 +99,9 @@ public class UserService {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new NotFoundException(NOT_FOUND_USER)
         );
+        if(userRepository.findByNickname(updateUserInfoRequestDto.nickname()).isPresent()){
+            throw new BadRequestException(DUPLICATED_NICKNAME);
+        }
         user.update(updateUserInfoRequestDto);
         fieldRepository.deleteByUserId(userId);
         List<Field> fieldList = updateUserInfoRequestDto.field().stream()
